fix: RFC 6386 compliance, dead code removal, and encoding quality improvements

+++ token.go
@@ -35,7 +35,10 @@ const (
 	PlaneY1SansY2 = 3 // Y blocks in B_PRED mode (no Y2)
 )
 
-// Legacy constants for backward compatibility (deprecated, use Plane* instead)
+// Legacy constants kept for backward compatibility.
+//
+// Deprecated: Use the Plane* constants instead. These aliases will be
+// removed in the next major version.
 const (
 	BlockTypeDCY  = PlaneY1WithY2 // Deprecated: use PlaneY1WithY2
 	BlockTypeACY  = PlaneY2       // Deprecated: use PlaneY2
@@ -1010,14 +1013,14 @@ func (h *CoeffHistogram) computeProbsForBand(baseProbs, updated *[4][8][3][11]ui
 
 // computeSingleProb computes a single updated probability from histogram counts.
 func (h *CoeffHistogram) computeSingleProb(baseProb uint8, i, j, k, l int) uint8 {
-	falseCount := h.counts[i][j][k][l][0]
-	trueCount := h.counts[i][j][k][l][1]
+	falseCount := uint64(h.counts[i][j][k][l][0])
+	trueCount := uint64(h.counts[i][j][k][l][1])
 	total := falseCount + trueCount
 
 	if total < 16 {
 		return baseProb
 	}
-	return clampProb((falseCount * 256) / total)
+	return clampProb(uint32(falseCount * 256 / total))
 }
 
 // clampProb clamps a probability value to the valid range [1, 255].
